Cinemachine Package Import, State Machine Implementation into Sprites
Imported the Cinemachine Packages from Unity Registry, so we can follow our player without having to code some egregious war crime. On top of that, a state machine had been made and correctly implemented so that whenever the sprite moves up it stops up, and works with all the other 3 directions (left, right, down.)

The cinemachine camera has also been implemented, whenver we test the camera now follows the player wherever they go. We can add stuff to this camera later, to prevent the camera from moving out of bounds when we make the map.