Combat Scene Set Up Pt. 1
Combat scene has been set up, but is still very rough. However, enemy targeting works, an arrow will highlight if selected and return back to the menu, but won't do any actions right now. You can also run away and it sends you back to the overworld scene, however it sends you back exactly to where you started. Will need to fix that to see if we can save the coordinates of the last combat initiated and also delete the Bat that initiated the combat with a player.