Tilemap, Map,  Camera Boundaries
Set up the tilemap, created the map from said tilemap with a rule tileset and then some, set up a boundary for the camera so that it doesn't cross into the blue space (Which is now black, which is fine.). Good news is that we're not using the background! We have a floor tileset!

All that's left from here now is to create the combat, placing the bats and detecting collisions, maybe more.