Turn-Based Combat, Collision detection, Prefabs and more!
This entire push is mainly focused on the implementation of Turn-Based Combat, which works and starts/completes matches based on if you won or lost.
It is a major feature implementation, and theres still more to go.

Collision detection works, so that when you collide with a bat it transports you to combat. It does not however destroy the enemy. Luckily, the bat just flies in place.

Prefabs have been made for the combat editor, but also for the overworld bats so we can just slide them into the world and then give them a level to load. In the future, perhaps we can load with an array and have it generate a random number within the array's index so that the battle chosen is actually "Random" instead of fixed like made here.

Next push will have more scenes and bats, as well as the Menu improvement.