Dragon Boss Fight, Main Menu Music
Finished setting up the boss fight complete with its own unique dialogue from start to finish, playtested it with no issues.
Made the big bat transition to that boss fight.

Also added main menu music, because  well I wanted to. Also finished giving credit to SwaggyWalkman for his Itch.

This is it for the string of updates, for now!